moving data directory
